roachtest: fix mixed cluster and node versions for tenant span stats
Previously, if the cluster and node versions were mixed (i.e. not the same), we would always expect the test to fail.
However, if both the cluster version and all node versions were at (or above) the "ToVersion" (v23.1.0), the test should
still succeed as we are still able to request span statistics using the new proto payload.

Release note: None