opt: handle passthrough exprs in SimplifyPartialIndexProjections
The `SimplifyPartialIndexProjections` normalization rule replaces partial index
expressions with the constant 'false' in update statements that do not change
any of the columns used in the partial index. This is an optimization to avoid
superfluously writing unchanged KVs to the partial index.

The rule assumed that all partial index expressions were projection columns. In
the case of a partial index expression `WHERE b` using a boolean column `b`,
however, the partial index expression could be a passthrough column. In such a
case the rule would apply recursively, running into the optimizer's stack depth
limit.

This commit adds support for passthrough partial index columns to the rule,
fixing the infinite recursion.

Fixes: #106663

Release note: None