sqlproxyccl: deflake TestDeleteTenant
There is a small race condition in the tenant directory cache.

1. If ReportFailure is called, the tenant dir rebuilds the cache by
	 listing the available pods.
2. There is a watch attempting to push events to the cache. If an event
	 was generated before the most recent list, the watch event can roll
	 back cache state.

This race is okay in production because when the sql proxy tries to dial
the stale pod, the dial will fail and the proxy will refresh the list
again.

This change adjusts the test to work like production. If the list is
stale, ReportFailure is retried to refresh the list of pods.

Fixes: #86077